fix: Switch NTP logic

+++ src/display/core/process/GrindProcess.h
@@ -52,7 +52,7 @@ class GrindProcess : public Process {
     }
 
     double getNewDelayTime() {
-        double newDelay = grindDelay + volumetricRateCalculator.getOvershootAdjustMillis(double(grindVolume), currentVolume);
+        double newDelay = grindDelay + volumetricRateCalculator.getOvershootAdjustMillis(grindVolume, currentVolume);
         ESP_LOGI("GrindProcess", "Setting new delay time - Old: %2f, Expected Volume: %f, Actual Volume: %2f, New Delay: %f",
                  grindDelay, grindVolume, currentVolume, newDelay);
         newDelay = std::clamp(newDelay, 0.0, PREDICTIVE_TIME);
